created a custom hook for tabs
